refactor: 패키지 스트럭쳐를 레이어에 따라 구성
