--- src/main/java/Repo.java
@@ -1,2 +1,0 @@
-package PACKAGE_NAME;public class Repo {
-}
