+++ src/main/java/subway/controller/PageController.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
