+++ src/main/java/subway/dao/LineDao.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.dao;
 
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -7,6 +7,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import subway.exception.DuplicateException;
+import subway.domain.Line;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -15,71 +16,77 @@ import java.sql.Statement;
 import java.util.List;
 import java.util.Optional;
 
-import static subway.station.StationQuery.*;
+import static subway.query.LineQuery.*;
 
 @Repository
-public class StationDao {
+public class LineDao {
     private final JdbcTemplate jdbcTemplate;
 
-    public StationDao(JdbcTemplate jdbcTemplate) {
+    public LineDao(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public Station save(Station station) {
-        if (hasDuplicateName(station.getName())) {
+    public Line save(Line line) {
+        if (hasDuplicateName(line.getName())) {
             throw new DuplicateException();
         }
-
-        return insertAtDB(station);
+        return insertAtDB(line);
     }
 
-    public List<Station> findAll() {
-        List<Station> stations = jdbcTemplate.query(selectAllQuery, new StationMapper());
-        return stations;
+    public void update(Line line) {
+        jdbcTemplate.update(updateQuery, line.getName(), line.getColor(), line.getId());
     }
 
-    public Optional<Station> findById(Long id) {
+    public Optional<Line> findById(Long id) {
         try {
-            return Optional.ofNullable(jdbcTemplate.queryForObject(selectByIdQuery, new StationMapper(), id));
+            return Optional.ofNullable(jdbcTemplate.queryForObject(selectByIdQuery, new LineMapper(), id));
         } catch (EmptyResultDataAccessException e) {
             return Optional.empty();
         }
+    }
+
+    public List<Line> findAll() {
+        List<Line> lines = jdbcTemplate.query(selectAllQuery, new LineMapper());
+        return lines;
     }
 
     public void deleteById(Long id) {
         jdbcTemplate.update(deleteByIdQuery, id);
     }
 
-    private boolean hasDuplicateName(String name) {
+    public boolean hasDuplicateName(String name) {
         return jdbcTemplate.queryForObject(countByNameQuery, int.class, name) != 0;
     }
 
-    private Station insertAtDB(Station station) {
+    private Line insertAtDB(Line line) {
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
         jdbcTemplate.update(con -> {
             PreparedStatement psmt = con.prepareStatement(
                     insertQuery,
                     Statement.RETURN_GENERATED_KEYS);
-            psmt.setString(1, station.getName());
+            psmt.setString(1, line.getName());
+            psmt.setString(2, line.getColor());
             return psmt;
         }, keyHolder);
 
         Long id = (Long) keyHolder.getKey();
 
-        return new Station(
+        return new Line(
                 id,
-                station.getName()
+                line.getName(),
+                line.getColor()
         );
     }
 
-    private final static class StationMapper implements RowMapper<Station> {
+    private final static class LineMapper implements RowMapper<Line> {
         @Override
-        public Station mapRow(ResultSet rs, int rowNum) throws SQLException {
+        public Line mapRow(ResultSet rs, int rowNum) throws SQLException {
             Long id = rs.getLong("id");
             String name = rs.getString("name");
+            String color = rs.getString("color");
 
-            return new Station(id, name);
+            return new Line(id, name, color);
         }
     }
 }
