+++ src/main/java/subway/exception/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
