+++ src/main/java/subway/response/LineResponse.java
@@ -1,6 +1,6 @@
-package subway.line;
+package subway.response;
 
-import subway.station.StationResponse;
+import subway.response.StationResponse;
 
 import java.util.List;
