+++ src/test/java/subway/line/LineDaoTest.java
@@ -7,6 +7,8 @@ import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabas
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import subway.dao.LineDao;
+import subway.domain.Line;
 
 import java.util.Optional;
